+++ config/prod.exs
@@ -24,8 +24,7 @@ config :last_crusader,
       }
     ]
   },
-  # 15 minutes
-  webmention_delai_ms: 900_000,
+  webmention_nb_tries: 15,
   port: 8080
 
 config :logger, :console,