+++ config/test.exs
@@ -25,8 +25,7 @@ config :last_crusader,
       }
     ]
   },
-  # 15 minutes
-  webmention_delai_ms: 900_000,
+  webmention_nb_tries: 15,
   port: 4002
 
 config :tesla, adapter: Tesla.Mock